更新微信accesstoken机制

+++ src/main/java/com/kl/parkLine/json/WxAccessTokenResult.java
@@ -15,5 +15,5 @@ public class WxAccessTokenResult extends WxResultBase
     private String accessToken;  
     
     @JSONField(name="expires_in")
-    private Long expiresIn;  
+    private Integer expiresIn;  
 }
